until 9-5.

+++ app/src/Controller/AnswersController.php
@@ -34,7 +34,7 @@ class AnswersController extends AppController
             return $this->redirect(['controller' => 'Question',
                                     'action' => 'view', $answer->question_id]);
         }
-        $answer->user_id = 1; // @TODO ユーザ管理濃機能実装時に修正する。
+        $answer->user_id = $this->Auth->user('id'); // @TODO ユーザ管理濃機能実装時に修正する。
         if($this->Answers->save($answer)){
             $this->Flash->success('回答を投稿しました。');
         }else{
@@ -55,6 +55,10 @@ class AnswersController extends AppController
         $answer = $this->Answers->get($id);
         $questionId = $answer->question_id;
 
+        if($answer->user_id !== $this->Auth->user('id')){
+            $this->Flash->error('他のユーザーの回答を削除することはできません。');
+            return $this->redirect->
+        }
         if($this->Answers->delete($answer)){
             $this->Flash->success('回答を削除しました。');
         }else{
