+++ app/src/Controller/QuestionsController.php
@@ -22,7 +22,7 @@ class QuestionsController extends AppController
     public function index()
     {
         // ここに処理を書いていく。
-        $questions = $this->paginate($this->Questions->findQuestionWithAnsweredCount(), [
+        $questions = $this->paginate($this->Questions->findQuestionWithAnsweredCount()->Contain(['Users']), [
             'order' => ['Questions.id' => 'DESC']
         ]);
         $this->set(compact('questions'));
@@ -42,7 +42,7 @@ class QuestionsController extends AppController
             if($this->request->is('post')){
                 $question = $this->Questions->patchEntity($question, $this->request->getData());
             }
-            $question->user_id = 1; // @TODOユーザ管理機能実装時に修正する。
+            $question->user_id = $this->Auth->user('id'); // @TODOユーザ管理機能実装時に修正する。
             if($this->Questions->save($question)){
                 $this->Flash->success('質問を投稿しました。');
 
@@ -54,11 +54,12 @@ class QuestionsController extends AppController
     }
     public function view(int $id)
     {
-        $question = $this->Questions->get($id);
+        $question = $this->Questions->get($id,['contain'=>['Users']]);
         $answers = $this
                  ->Answers
                  ->find()
                  ->where(['Answers.question_id' => $id])
+                 ->contain(['Users'])
                  ->orderAsc('Answers.id')
                  ->all();
 
@@ -75,8 +76,10 @@ class QuestionsController extends AppController
         $this->request->allowMethod(['post']);
 
         $question = $this->Questions->get($id);
-
-        // @TODO 質問を削除できるのは質問投稿者のみとする
+        if($question->user_id !== $this->Auth->user('id')){
+            $this->Flash->error('他のユーザの質問を削除することはできません。');
+            return $this->redirect(['action'=>'index']);
+        }
 
         if($this->Questions->delete($question)){
             $this->Flash->success('質問を削除しました。');
